requirements.txt
-e git+https://github.com/kevinastone/django-api-rest-and-angular.git@b0a1ea487c6d7028e491e7558a7a0cfd2243d2ec#egg=API_Example
bootstrap-admin==0.3.7.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-extensions==1.8.1
django-oauth-toolkit==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.6.3
idna==2.5
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pytz==2017.2
requests==2.18.3
six==1.10.0
sorl-thumbnail==12.3
urllib3==1.22
Werkzeug==0.12.2
